refactor(host-internal): LSP 编排器支持多 provider 会话
新增 LspProviderSession 与 LspOrchestrator，LspService 改为门面；TypeScript 端到端行为保持不变。
