+++ packages/host-internal/package.json
@@ -70,7 +70,7 @@
     "build": "npm --prefix ../agent-core run build && npm run build:tsc",
     "build:tsc": "tsc -p tsconfig.json",
     "test:workspace-file-references": "npm run build && node --test dist/workspace-file-references.test.js",
-    "test:lsp": "npm run build && node --test dist/lsp/providers.test.js dist/lsp/service-probe.test.js dist/lsp/write-append.test.js",
+    "test:lsp": "npm run build && node --test dist/lsp/providers.test.js dist/lsp/resolve-server.test.js dist/lsp/orchestrator.test.js dist/lsp/service-probe.test.js dist/lsp/write-append.test.js",
     "smoke:lsp": "npm run build && node dist/smoke/lsp-typescript-smoke.js",
     "typecheck": "npm --prefix ../agent-core run build && tsc -p tsconfig.json --noEmit",
     "smoke:openai-models-parse": "npm run build && node dist/smoke/openai-models-parse-smoke.js",
