+++ packages/host-internal/src/lsp/index.ts
@@ -2,8 +2,10 @@ export * from './config.js';
 export * from './connection.js';
 export * from './document-store.js';
 export * from './errors.js';
+export * from './orchestrator.js';
 export * from './parse-change.js';
 export * from './paths.js';
+export * from './provider-session.js';
 export * from './providers.js';
 export * from './resolve-server.js';
 export * from './service.js';
